@@ -36,7 +36,7 @@
 		<!--This will get replaced by Build System-->
 		<Version>0.0.1</Version>
 
-		<LangVersion>12.0</LangVersion>
+		<LangVersion>14.0</LangVersion>
 		<ImplicitUsings>enable</ImplicitUsings>
 
 		<!--Subject to change by project that need to be packed-->
