@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net9.0</TargetFramework>
+		<TargetFramework>net10.0</TargetFramework>
 
 		<!--Be able to browse generated files-->
 		<EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>
