@@ -3,58 +3,58 @@
     [TestFixture]
     public class SpanParserHelperTests
     {
-        private static IEnumerable<TestCaseData> ValidValuesFor1() => new[]
-        {
-            new TestCaseData(@"", @"", ';'),
-            new TestCaseData(@";", @";", ';'),
-            new TestCaseData(@"\;", @";", ';'),
-            new TestCaseData(@"XYZ\", @"XYZ\", ';'),
-            new TestCaseData(@"ABC", @"ABC", ';'),
-            new TestCaseData(@"ABC\\;", @"ABC\\;", ';'),
-            new TestCaseData(@"ABC\\\;", @"ABC\\;", ';'),
+        private static IEnumerable<TCD> ValidValuesFor1() =>
+        [
+            new TCD(@"", @"", ';'),
+            new TCD(@";", @";", ';'),
+            new TCD(@"\;", @";", ';'),
+            new TCD(@"XYZ\", @"XYZ\", ';'),
+            new TCD(@"ABC", @"ABC", ';'),
+            new TCD(@"ABC\\;", @"ABC\\;", ';'),
+            new TCD(@"ABC\\\;", @"ABC\\;", ';'),
 
-            new TestCaseData(@"ABC\\;DEF", @"ABC\\;DEF", ';'),
-            new TestCaseData(@"ABC\\\;DEF", @"ABC\\;DEF", ';'),
+            new TCD(@"ABC\\;DEF", @"ABC\\;DEF", ';'),
+            new TCD(@"ABC\\\;DEF", @"ABC\\;DEF", ';'),
 
-            new TestCaseData(@"ABC\\;DE\F", @"ABC\\;DE\F", ';'),
-            new TestCaseData(@"ABC\\\;DE\F", @"ABC\\;DE\F", ';'),
-            new TestCaseData(@"ABC\\\;DE\F\;", @"ABC\\;DE\F;", ';'),
-            new TestCaseData(@"ABC\\\;DE\F\;", @"ABC\\;DE\F\;", '\\'),
+            new TCD(@"ABC\\;DE\F", @"ABC\\;DE\F", ';'),
+            new TCD(@"ABC\\\;DE\F", @"ABC\\;DE\F", ';'),
+            new TCD(@"ABC\\\;DE\F\;", @"ABC\\;DE\F;", ';'),
+            new TCD(@"ABC\\\;DE\F\;", @"ABC\\;DE\F\;", '\\'),
 
-            new TestCaseData(@"\ABC\", @"\ABC\", ';'),
-            new TestCaseData(@"ABC\", @"ABC\", ';'),
-            new TestCaseData(@"\ABC", @"\ABC", ';'),
+            new TCD(@"\ABC\", @"\ABC\", ';'),
+            new TCD(@"ABC\", @"ABC\", ';'),
+            new TCD(@"\ABC", @"\ABC", ';'),
 
-            new TestCaseData(@";ABC;", @";ABC;", ';'),
-            new TestCaseData(@"ABC;", @"ABC;", ';'),
-            new TestCaseData(@";ABC", @";ABC", ';'),
+            new TCD(@";ABC;", @";ABC;", ';'),
+            new TCD(@"ABC;", @"ABC;", ';'),
+            new TCD(@";ABC", @";ABC", ';'),
 
-            new TestCaseData(@"\;ABC\;", @";ABC;", ';'),
-            new TestCaseData(@"ABC\;", @"ABC;", ';'),
-            new TestCaseData(@"\;ABC", @";ABC", ';'),
+            new TCD(@"\;ABC\;", @";ABC;", ';'),
+            new TCD(@"ABC\;", @"ABC;", ';'),
+            new TCD(@"\;ABC", @";ABC", ';'),
 
-            new TestCaseData(@"\\ABC\\", @"\\ABC\\", ';'),
-            new TestCaseData(@"ABC\\", @"ABC\\", ';'),
-            new TestCaseData(@"\\ABC", @"\\ABC", ';'),
+            new TCD(@"\\ABC\\", @"\\ABC\\", ';'),
+            new TCD(@"ABC\\", @"ABC\\", ';'),
+            new TCD(@"\\ABC", @"\\ABC", ';'),
 
-            new TestCaseData(@"\A\B\C", @"\A\B\C", ';'),
-            new TestCaseData(@"\A\B\C", @"A\B\C", 'A'),
-            new TestCaseData(@"\A\B\C", @"\AB\C", 'B'),
-            new TestCaseData(@"\A\B\C", @"\A\BC", 'C'),
+            new TCD(@"\A\B\C", @"\A\B\C", ';'),
+            new TCD(@"\A\B\C", @"A\B\C", 'A'),
+            new TCD(@"\A\B\C", @"\AB\C", 'B'),
+            new TCD(@"\A\B\C", @"\A\BC", 'C'),
 
-            new TestCaseData(@"ABC\;", @"ABC;", ';'),
-            new TestCaseData(@"\\ABC\\\;", @"\\ABC\\;", ';'),
+            new TCD(@"ABC\;", @"ABC;", ';'),
+            new TCD(@"\\ABC\\\;", @"\\ABC\\;", ';'),
 
-            new TestCaseData(@"\\ABC\;", @"\ABC\;", '\\'),
-            new TestCaseData(@"\\ABC\\;", @"\ABC\;", '\\'),
-            new TestCaseData(@"\\ABC\\\;", @"\ABC\\;", '\\'),
-            new TestCaseData(@"\\ABC\\\\;", @"\ABC\\;", '\\'),
+            new TCD(@"\\ABC\;", @"\ABC\;", '\\'),
+            new TCD(@"\\ABC\\;", @"\ABC\;", '\\'),
+            new TCD(@"\\ABC\\\;", @"\ABC\\;", '\\'),
+            new TCD(@"\\ABC\\\\;", @"\ABC\\;", '\\'),
 
-            new TestCaseData(@"\\ABC\", @"\ABC\", '\\'),
-            new TestCaseData(@"\\ABC\\", @"\ABC\", '\\'),
-            new TestCaseData(@"\\ABC\\\", @"\ABC\\", '\\'),
-            new TestCaseData(@"\\ABC\\\\", @"\ABC\\", '\\'),
-        };
+            new TCD(@"\\ABC\", @"\ABC\", '\\'),
+            new TCD(@"\\ABC\\", @"\ABC\", '\\'),
+            new TCD(@"\\ABC\\\", @"\ABC\\", '\\'),
+            new TCD(@"\\ABC\\\\", @"\ABC\\", '\\'),
+        ];
 
         [TestCaseSource(nameof(ValidValuesFor1))]
         public void UnescapeCharacter1Tests(string input, string expectedOutput, char character)
@@ -65,74 +65,74 @@
             Assert.That(actual, Is.EqualTo(expectedOutput));
         }
 
-        private static IEnumerable<TestCaseData> ValidValuesFor2() => ValidValuesFor1().Select(
-            tcd => new TestCaseData(tcd.Arguments[0], tcd.Arguments[1], tcd.Arguments[2], '=')
-        ).Concat(new[]
-        {
-            new TestCaseData(@"", @"", ';', ','),
-            new TestCaseData(@";,", @";,", ';', ','),
-            new TestCaseData(@"\;\,", @";,", ';', ','),
-            new TestCaseData(@"\,\;\,", @",;,", ';', ','),
-            new TestCaseData(@"\;\,\;\,", @";,;,", ';', ','),
-            new TestCaseData(@"\;\,\;\,\;", @";,;,;", ';', ','),
-            new TestCaseData(@"XYZ\ ,", @"XYZ\ ,", ';', ','),
-            new TestCaseData(@"ABC\ ;", @"ABC\ ;", ';', ','),
-            new TestCaseData(@"ABC\\;", @"ABC\\;", ';', ','),
-            new TestCaseData(@"ABC\\,", @"ABC\\,", ';', ','),
-            new TestCaseData(@"ABC\\\;", @"ABC\\;", ';', ','),
-            new TestCaseData(@"ABC\\\,", @"ABC\\,", ';', ','),
+        private static IEnumerable<TCD> ValidValuesFor2() => ValidValuesFor1().Select(
+            tcd => new TCD(tcd.Arguments[0], tcd.Arguments[1], tcd.Arguments[2], '=')
+        ).Concat(
+        [
+            new TCD(@"", @"", ';', ','),
+            new TCD(@";,", @";,", ';', ','),
+            new TCD(@"\;\,", @";,", ';', ','),
+            new TCD(@"\,\;\,", @",;,", ';', ','),
+            new TCD(@"\;\,\;\,", @";,;,", ';', ','),
+            new TCD(@"\;\,\;\,\;", @";,;,;", ';', ','),
+            new TCD(@"XYZ\ ,", @"XYZ\ ,", ';', ','),
+            new TCD(@"ABC\ ;", @"ABC\ ;", ';', ','),
+            new TCD(@"ABC\\;", @"ABC\\;", ';', ','),
+            new TCD(@"ABC\\,", @"ABC\\,", ';', ','),
+            new TCD(@"ABC\\\;", @"ABC\\;", ';', ','),
+            new TCD(@"ABC\\\,", @"ABC\\,", ';', ','),
 
-            new TestCaseData(@"ABC\\,;DEF", @"ABC\\,;DEF", ';', ','),
-            new TestCaseData(@"ABC\\\,;DEF", @"ABC\\,;DEF", ';', ','),
+            new TCD(@"ABC\\,;DEF", @"ABC\\,;DEF", ';', ','),
+            new TCD(@"ABC\\\,;DEF", @"ABC\\,;DEF", ';', ','),
 
-            new TestCaseData(@"ABC\\,;DE\,F", @"ABC\\,;DE,F", ';', ','),
-            new TestCaseData(@"ABC\\\,;DE\,F", @"ABC\\,;DE,F", ';', ','),
-            new TestCaseData(@"ABC\\\;DE\F\,;", @"ABC\\;DE\F,;", ';', ','),
-            new TestCaseData(@"ABC\\\;DE\F,\,", @"ABC\\;DE\F,,", '\\', ','),
+            new TCD(@"ABC\\,;DE\,F", @"ABC\\,;DE,F", ';', ','),
+            new TCD(@"ABC\\\,;DE\,F", @"ABC\\,;DE,F", ';', ','),
+            new TCD(@"ABC\\\;DE\F\,;", @"ABC\\;DE\F,;", ';', ','),
+            new TCD(@"ABC\\\;DE\F,\,", @"ABC\\;DE\F,,", '\\', ','),
 
-            new TestCaseData(@"\ABC\", @"\ABC\", ';', ','),
-            new TestCaseData(@"ABC\", @"ABC\", ';', ','),
-            new TestCaseData(@"\ABC", @"\ABC", ';', ','),
+            new TCD(@"\ABC\", @"\ABC\", ';', ','),
+            new TCD(@"ABC\", @"ABC\", ';', ','),
+            new TCD(@"\ABC", @"\ABC", ';', ','),
 
-            new TestCaseData(@";,ABC,;", @";,ABC,;", ';', ','),
-            new TestCaseData(@"ABC,;", @"ABC,;", ';', ','),
-            new TestCaseData(@",;ABC", @",;ABC", ';', ','),
+            new TCD(@";,ABC,;", @";,ABC,;", ';', ','),
+            new TCD(@"ABC,;", @"ABC,;", ';', ','),
+            new TCD(@",;ABC", @",;ABC", ';', ','),
 
-            new TestCaseData(@",;ABC;,", @",;ABC;,", ';', ','),
-            new TestCaseData(@"ABC;,", @"ABC;,", ';', ','),
-            new TestCaseData(@";,ABC", @";,ABC", ';', ','),
-
-
-            new TestCaseData(@"\;\,ABC\,\;", @";,ABC,;", ';', ','),
-            new TestCaseData(@"ABC\,\;", @"ABC,;", ';', ','),
-            new TestCaseData(@"\,\;ABC", @",;ABC", ';', ','),
-
-            new TestCaseData(@"\,\;ABC\;\,", @",;ABC;,", ';', ','),
-            new TestCaseData(@"ABC\;\,", @"ABC;,", ';', ','),
-            new TestCaseData(@"\;\,ABC", @";,ABC", ';', ','),
-
-            new TestCaseData(@"\\ABC\\", @"\\ABC\\", ';', ','),
-            new TestCaseData(@"ABC\\", @"ABC\\", ';', ','),
-            new TestCaseData(@"\\ABC", @"\\ABC", ';', ','),
-
-            new TestCaseData(@"\\ABC\\", @"\ABC\", '\\', ','),
-            new TestCaseData(@"ABC\\", @"ABC\", '\\', ','),
-            new TestCaseData(@"\\ABC", @"\ABC", '\\', ','),
-
-            new TestCaseData(@"\A\B\C", @"\A\B\C", ';', ','),
-            new TestCaseData(@"\A\B\C", @"AB\C", 'A', 'B'),
-            new TestCaseData(@"\A\B\C", @"\ABC", 'C', 'B'),
-            new TestCaseData(@"\A\B\C", @"A\BC", 'A', 'C'),
+            new TCD(@",;ABC;,", @",;ABC;,", ';', ','),
+            new TCD(@"ABC;,", @"ABC;,", ';', ','),
+            new TCD(@";,ABC", @";,ABC", ';', ','),
 
 
-            new TestCaseData(@"\,ABC\;", @",ABC;", ';', ','),
-            new TestCaseData(@"\\,ABC\\;", @"\\,ABC\\;", ';', ','),
-            new TestCaseData(@"\\\,ABC\\\;", @"\\,ABC\\;", ';', ','),
-            new TestCaseData(@"\\\,ABC\\\;", @"\,ABC\\;", '\\', ','),
-            new TestCaseData(@"\\\,ABC\\\;", @"\\,ABC\;", '\\', ';'),
+            new TCD(@"\;\,ABC\,\;", @";,ABC,;", ';', ','),
+            new TCD(@"ABC\,\;", @"ABC,;", ';', ','),
+            new TCD(@"\,\;ABC", @",;ABC", ';', ','),
 
-            new TestCaseData(@"\\ABC\\\;", @"\ABC\;", '\\', ';'),
-        });
+            new TCD(@"\,\;ABC\;\,", @",;ABC;,", ';', ','),
+            new TCD(@"ABC\;\,", @"ABC;,", ';', ','),
+            new TCD(@"\;\,ABC", @";,ABC", ';', ','),
+
+            new TCD(@"\\ABC\\", @"\\ABC\\", ';', ','),
+            new TCD(@"ABC\\", @"ABC\\", ';', ','),
+            new TCD(@"\\ABC", @"\\ABC", ';', ','),
+
+            new TCD(@"\\ABC\\", @"\ABC\", '\\', ','),
+            new TCD(@"ABC\\", @"ABC\", '\\', ','),
+            new TCD(@"\\ABC", @"\ABC", '\\', ','),
+
+            new TCD(@"\A\B\C", @"\A\B\C", ';', ','),
+            new TCD(@"\A\B\C", @"AB\C", 'A', 'B'),
+            new TCD(@"\A\B\C", @"\ABC", 'C', 'B'),
+            new TCD(@"\A\B\C", @"A\BC", 'A', 'C'),
+
+
+            new TCD(@"\,ABC\;", @",ABC;", ';', ','),
+            new TCD(@"\\,ABC\\;", @"\\,ABC\\;", ';', ','),
+            new TCD(@"\\\,ABC\\\;", @"\\,ABC\\;", ';', ','),
+            new TCD(@"\\\,ABC\\\;", @"\,ABC\\;", '\\', ','),
+            new TCD(@"\\\,ABC\\\;", @"\\,ABC\;", '\\', ';'),
+
+            new TCD(@"\\ABC\\\;", @"\ABC\;", '\\', ';'),
+        ]);
 
 
         [TestCaseSource(nameof(ValidValuesFor2))]
